tests: fold feature-search tests into test_fbtree.cpp
Merge test_fbtree_feature_search.cpp into test_fbtree.cpp, matching the
one-test-file-per-module convention in src/unit. The FeatureSearchTest
suite name is unchanged, so filtering and output grouping still work.
The local FEATURE_SIZE/FEATURE_ROW_SIZE/FEATURE_BIAS definitions are
dropped in favor of fbtree_internal.h (already included), and a stale
comment referencing fbtree_ordered_index.c is corrected to fbtree.c.
